+++ ot_simple_rest/handlers/eva/timelines.py
@@ -29,17 +29,19 @@ class GetTimelines(tornado.web.RequestHandler):
     """
 
     def initialize(self, mem_conf: Dict, static_conf: Dict):
+        self.loader = TimelinesLoader(mem_conf, static_conf)
         self.builder = TimelinesBuilder()
-        self.loader = TimelinesLoader(mem_conf, static_conf, self.builder.BIGGEST_INTERVAL)
 
     async def get(self):
         params = self.request.query_arguments
         cid = params.get('cid')[0].decode()
         try:
-            data = self.loader.load_data(cid)
-            timelines = self.builder.get_all_timelines(data)
+            data, fresh_time = self.loader.load_data(cid)  # fresh_time indicates last time interval in all timelines
+            timelines = self.builder.get_all_timelines(data, fresh_time)
         except tornado.web.HTTPError as e:
             return self.write(json.dumps({'status': 'failed', 'error': e}, default=str))
+        except KeyError:
+            return self.write(json.dumps({'status': 'failed', 'error': "'_time' column is missing"}, default=str))
         except Exception as e:
             return self.write(json.dumps({'status': 'failed', 'error': f'{e} cid {cid}'}, default=str))
         self.write(json.dumps(timelines))